ENH: Spin off ChunkLocations functionality from CreateTensorFlowDataset
